requirements/requirements.txt
requests
aws-xray-sdk
pyjwt